Move vm code under src/worker
