+++ benchmarks/clickhouse/e2e.kafka.benchmark.ts
@@ -5,8 +5,8 @@ import { startPluginsServer } from '../../src/server'
 import { LogLevel, PluginsServerConfig, Queue } from '../../src/types'
 import { PluginsServer } from '../../src/types'
 import { delay, UUIDT } from '../../src/utils'
-import { createPosthog, DummyPostHog } from '../../src/vm/extensions/posthog'
 import { makePiscina } from '../../src/worker/piscina'
+import { createPosthog, DummyPostHog } from '../../src/worker/vm/extensions/posthog'
 import { resetTestDatabaseClickhouse } from '../../tests/helpers/clickhouse'
 import { resetKafka } from '../../tests/helpers/kafka'
 import { pluginConfig39 } from '../../tests/helpers/plugins'
