+++ src/worker/vm/extensions/console.ts
