+++ src/worker/vm/lazy.ts
@@ -1,6 +1,6 @@
-import { clearError, processError } from '../error'
-import { status } from '../status'
-import { PluginConfig, PluginConfigVMReponse, PluginsServer, PluginTask } from '../types'
+import { clearError, processError } from '../../error'
+import { status } from '../../status'
+import { PluginConfig, PluginConfigVMReponse, PluginsServer, PluginTask } from '../../types'
 import { createPluginConfigVM } from './vm'
 
 export class LazyPluginVM {
