+++ src/worker/vm/transforms/common.ts
@@ -1,6 +1,6 @@
 import { PluginObj } from '@babel/core'
 import * as types from '@babel/types'
 
-import { PluginsServer } from '../../types'
+import { PluginsServer } from '../../../types'
 
 export type PluginGen = (server: PluginsServer) => (param: { types: typeof types }) => PluginObj
