+++ src/worker/vm/transforms/loop-timeout.ts
