+++ src/worker/vm/vm.ts
@@ -2,7 +2,7 @@ import { randomBytes } from 'crypto'
 import fetch from 'node-fetch'
 import { VM } from 'vm2'
 
-import { PluginConfig, PluginConfigVMReponse, PluginsServer } from '../types'
+import { PluginConfig, PluginConfigVMReponse, PluginsServer } from '../../types'
 import { createCache } from './extensions/cache'
 import { createConsole } from './extensions/console'
 import { createGoogle } from './extensions/google'
