+++ tests/postgres/e2e.test.ts
@@ -4,8 +4,8 @@ import { startPluginsServer } from '../../src/server'
 import { LogLevel } from '../../src/types'
 import { PluginsServer } from '../../src/types'
 import { UUIDT } from '../../src/utils'
-import { createPosthog, DummyPostHog } from '../../src/vm/extensions/posthog'
 import { makePiscina } from '../../src/worker/piscina'
+import { createPosthog, DummyPostHog } from '../../src/worker/vm/extensions/posthog'
 import { pluginConfig39 } from '../helpers/plugins'
 import { resetTestDatabase } from '../helpers/sql'
 import { delayUntilEventIngested } from '../shared/process-event'
