+++ tests/postgres/e2e.timeout.test.ts
@@ -1,8 +1,8 @@
 import { startPluginsServer } from '../../src/server'
 import { LogLevel, PluginsServer } from '../../src/types'
 import { UUIDT } from '../../src/utils'
-import { createPosthog, DummyPostHog } from '../../src/vm/extensions/posthog'
 import { makePiscina } from '../../src/worker/piscina'
+import { createPosthog, DummyPostHog } from '../../src/worker/vm/extensions/posthog'
 import { pluginConfig39 } from '../helpers/plugins'
 import { resetTestDatabase } from '../helpers/sql'
 import { delayUntilEventIngested } from '../shared/process-event'
